Changes
Combined move and after_move, to have everything in one single function.

Create an amoeba_state class for easy understanding of function call.

Made the frame rendering faster by adding the bacteria to the grid. The quality might be low, but significantly faster.

Noticed a bug in the GUI so temporarily turned it off.